new format

+++ contains/contains.h
@@ -1,4 +1,4 @@
-int contains(int vect[],int size,int elem) {
+size_t contains(int vect[],size_t size,int elem) {
 	//Ecrire le code de la fonction
 	return -1;
 }
